Fix and refactor tests
