+++ dod/src/test/java/com/revolut/rxdata/dod/TestParamsDomain.kt
@@ -1,0 +1,4 @@
+package com.revolut.rxdata.dod
+
+typealias Params = Int
+typealias Domain = String
